lib/oelite: remove lots of dead code
Despite the vague comment above the creation of the task_depend table,
I think it, and all the associated code should be removed, since it is
both completely unused in practice and somewhat buggy.

In OEliteTask::get_taskdepends, we fetch two columns from the table,
but then hit the result with flatten_single_column_rows, despite the
code in runq.py obviously expecting to get a list of pairs. Also, if
we want to unconditionally throw an exception if someone tries to use
an as-yet-unimplemented feature, I think that would better done where
we first detect that someone tries to do do_foo[depends] = "bar:do_baz".