passwd classes: make them more robust
There's a bug in the current avahi recipe: The autoipd-passwd and
daemon-passwd files end up being packaged under the names avahi.1 and
avahi.2, but in the packages' respective FILES_* list they list each
others' files. This means that when one includes one but not the other
package, the expected user account is missing from the system.

Having to manually enumerate the PASSWD_FILES the same way the passwd
class does it is rather fragile. So instead of renaming the files to
${PN}.${i}, put each recipe's passwd files in the subdirectory ${PN} and
keep the file's basename. That makes the relevant FILES entries much
more predictable. We pass -n to cp to catch files with the same basename.

While at it, replace $cwd with . in image_passed.oeclass - avoiding
long absolute paths makes it easier to read the log output.

The initial /dev/null argument to the cat commands is just a standard
idiom to ensure that, in the unlikely case find doesn't actually find
any files, cat doesn't sit around waiting for stdin (which, I believe,
is normally already redirected from /dev/null, but it wouldn't be if
one is manually running the do_install script).