auto-package-utils.oeclass: also handle man pages for alternate invocation names
When a binary such as xz comes with a symlink unxz->xz, the
documentation usually also contains a corresponding symlink (or, as is
the case for bzip2, a file containing nothing but a '.so man1/bzip2.1'
directive, which amounts to the same thing). Make these files or
symlinks go into the doc package providing the real man page.