lib/oelite: replace datetime.datetime.now calls
datetime.datetime.now() returns time in the local timezone. Hence, it
reads /etc/localtime - and caches the result, so subsequent calls just
stat() that file. Still, getting a timestamp for delta computations
shouldn't involve any syscalls at all on a modern platform with a
vdso.

It turns out datetime.utcnow doesn't do any system calls (only a
gettimeofday() vdso call) and is about three times faster, but it still
has an overhead of around 350 nanoseconds per call, presumably because
of all the python code involved in building a full datetime.datetime
object.

That's quite wasteful, since we only want this for delta computations,
which means we end up having to apply the .total_seconds() method to the
difference. So a much better choice is time.time (overhead ~40
nanoseconds), which returns the gettimeofday() result as a (python)
float. We're around 2^31 seconds after the epoch, leaving 22 bits for
the fractional part, so we're not losing any precision this way.

Abstract the choice in oelite.util.now, which must return a float so it
can be used directly for delta computations, and use that throughout.