oelite/baker.py: slightly refactor error handling in baking loop
As preparation for the next steps towards parallel tasks, reduce the
amount of local state that needs to be updated in the error path -
we're going to move most of the logic in that loop to a separate file.

As a side-effect, we get to outdent the final loop one level.