oelite/baker.py: poll every task in the oven
To make best use of our resources, it is important that we note the
succesful completion of tasks as fast as possible. Otherwise, if we
have lots of pending synchronous tasks lined up (do_fetch and
friends), it may be a long time before we enter the "we're out of
tasks or room in the oven" branch. This can cause something like
glibc:do_configure to lie around.

So rather than just doing a single poll for the task we just started,
poll every currently baking task (all but the one we just started are
obviously guaranteed to be asynchronous).