oelite/task.py: open log files with CLOEXEC flag
Doing 'ls -l /proc/*/fd' during an oe bake, one sees that tasks
typically inherits file descriptors for completely unrelated tasks'
log files (as well as always one extra to its own). This is mostly a
cosmetic issue, but can cause some surprising and hard-to-debug
results if some process misbehaves, so we might as well ask the kernel
to close the irrelevant fds - that also reduces the cost of fork()ing
in the descendant processes slightly (since there's fewer kernel file
objects to do reference bookkeeping on).

As a small bonus, since we never do I/O to the logfile file object
itself, we only create the raw file descriptor, so Python doesn't
allocate a whole file object and an associated buffer.