classes/*: cache magic database
In runstrip(), we open, read and parse the system's magic database for
each and every file we process. That's quite wasteful; just in terms
of raw system calls this is something like:

  open("/etc/magic.mgc", O_RDONLY)        = -1 ENOENT (No such file or directory) <0.000009>
  stat("/etc/magic", {st_mode=S_IFREG|0644, st_size=111, ...}) = 0 <0.000009>
  open("/etc/magic", O_RDONLY)            = 9 <0.000008>
  fstat(9, {st_mode=S_IFREG|0644, st_size=111, ...}) = 0 <0.000006>
  mmap(NULL, 4096, PROT_READ|PROT_WRITE, MAP_PRIVATE|MAP_ANONYMOUS, -1, 0) = 0x7f4447b78000 <0.000008>
  read(9, "# Magic local data for file(1) c"..., 4096) = 111 <0.000009>
  read(9, "", 4096)                       = 0 <0.000006>
  close(9)                                = 0 <0.000007>
  munmap(0x7f4447b78000, 4096)            = 0 <0.000008>
  open("/usr/share/misc/magic.mgc", O_RDONLY) = 9 <0.000013>
  fstat(9, {st_mode=S_IFREG|0644, st_size=2614912, ...}) = 0 <0.000006>
  mmap(NULL, 2614912, PROT_READ|PROT_WRITE, MAP_PRIVATE, 9, 0) = 0x7f4437638000 <0.000008>
  close(9)                                = 0 <0.000006>
  ...
  munmap(0x7f4437638000, 2614912)         = 0 <0.000014>

and then there's actual database parsing and python runtime overhead.

We also use libmagic from a few other places, and in some of those, we
do rely on one opened+loaded filemagic object being able to parse
several files before being disposed of. So let's create a simple
wrapper class that caches such a filemagic object for us, with methods
that simply forward to that object.

One obvious restriction: The magic database to use has to be part of
the open() call just as the flags argument. But in our cases,
everybody passes magic.MAGIC_NONE for flags and use the default
database, so this should be ok. We could easily implement "lazy loading"
with an internal load_called flag, but since every user calls .load()
immediately after .open() we might as well do the loading immediately,
and not provide a .load() method.

In chrpath, we used to leak the database handle (magic.py is a thin
ctypes-based wrapper for libmagic.so, so resource management has to be
explicit). That left us with another 2.5MB of VM and an extra entry in
/proc/self/maps - during an oe bake world, we used to end up mapping
/usr/share/file/magic.mgc hundreds of times, costing over 700 MB of
VM. I've added a .close() call on the normal return path, but it no
longer matters much since we never really close the _single_ cached
handle.