oelite/meta/dict.py: use list instead of dict for override bookkeeping
$ python
Python 2.7.12 (default, Jul  1 2016, 15:12:24)
[GCC 5.4.0 20160609] on linux2
Type "help", "copyright", "credits" or "license" for more information.
>>> import sys
>>> x=dict()
>>> y=[1, 2, 3]
>>> sys.getsizeof(x)
280
>>> sys.getsizeof(y)
96

So by using a list (with a fixed internal mapping from '', '>', '<' to
indices) instead of dict for the __overrides, we save 184 bytes for
every var with an override. We end up having over 130000 of those, so
the net saving is more then 25 MiB (about 3% of our memory consumption)
for an oe bake world.