oelite/meta/meta.py: add REBUILDALL_SKIP to builtin_nohash
Recipes which do not set REBUILDALL_SKIP (by inheriting one of the few
classes where that is set) currently set the nohash flag for a
non-existing variable. That wastes a lot of memory (about 1.4 MiB). So
just make it a built-in rule that a variable with this name is nohash.