oelite/profiling.py: make @profile_calls more DWIMy
Currently, if one adds the @profile_calls decorator to an inner
function, one gets one line of output for every time the inner function
got defined (that is, once for every time the "def foobar()" line gets
executed by the Python interpreter). This is most likely not what one
expects, and certainly not how I meant it to work.

To make the profiling low overhead, we still want to be able to use the
function object as a dict key during execution. But when we're about to
write out the stats, do a pass over the profiled_functions dictionary
and collect the call stats for functions with the same (sourcefile,
lineno, function name) tuple.