oelite/baker.py: add --dryrun option
This makes it easier to do all the initial recipe parsing, hash
computation etc. without doing an actual build - this is especially
useful if one is doing some automated testing of a series of meta/core
commits, since one doesn't have to sit and stare at the terminal for a
few minutes waiting to say 'no'. It can also be useful in
non-interactive settings, for example setting up a cheap buildbot job to
do "oe bake world --dryrun" to check that all recipes actually parse,
without doing a full build.