oelite/dbutil.py: add query profiling infrastructure
I want to know which sql queries take the most time, so that we can
either add indexes, rewrite the queries or completely get rid of the
relevant sql code. Rather than adding ugly wrappers around each and
every .execute and .executemany call, create a simple wrapper for a
database cursor.

The __getattr__ method ensures that even if we end up calling some
method or accessing some attribute other than .execute and .executemany
this should DTRT (that is, forward to the actual cursor object).

When passing the default profile=False, the slowdown compared to this
code not being present is in the noise. When passing profile=True, we
take a performance hit of about 10-20%, so I definitely don't think this
should be enabled by default.