oelite/dbutil.py: capture full cost of queries
I was running "perf record oe bake world --dryrun" and got a somewhat
surprising result:

  51,43%  oe            libsqlite3.so.0.8.6
  36,09%  oe            python2.7
   9,13%  oe            libc-2.19.so
   1,61%  oe            libpthread-2.19.so
   0,81%  oe            _sqlite.so

But this wasn't consistent with a total runtime of about 150
seconds, with the sum of all query times only being about 16 seconds. So
how can we end up spending that much time in libsqlite? Turns out I
wasn't measuring the cost of the row-fetching, which often dwarfs the
time to do SELECT itself.

So we have to update our cursor wrapper to be iterable (so that it can
be used in a direct "for x in dbc.execute(...)"), and also provide
.fetchone and .fetchall methods, in all cases measuring the time spent
in the underlying call, and attributing that back to the most recent
query.

Pretty ugly, and it adds more overhead, but it really helps finding the
problematic queries and tables.