classes/sysvinit.oeclass: prepare for USE flag semantics change
Currently, setting MACHINE_USE_foo_sysvinit_stop = "0" means the
USE_foo_sysvinit_stop variable ends up being unset rather than having
the value "0". This is counter-intuitive, so we'll change that later. To
prepare for that while preserving back- and forward-compatibility, make
the documented way to disable the start/stop script to set the value to
the token False.

Inside do_install_sysvinit, we recognize anything that is False in the
eyes of python as well as the legacy string "0".

After that, anything other than a two-digit string is a user error, and
we don't do anybody any services by silently ignoring that.

Because of the huge semantic difference between "0" and "00", and
because False conveys the intention much better than any string value
would, people are strongly encouraged to update their recipes, distro
and machine configs etc. to use False instead of "0". Since the current
OE-lite lexer internally translates the token False to the string "0",
this can be safely be done with or without a meta/core containing this
patch.