classes/crontab.oeclass: only invoke do_install_crontab if USE_crontab
One cannot combine flag and override assignments, so the most "obvious"
thing

  do_install[postfuncs]:>USE_crontab += " do_install_crontab"

doesn't work. Instead, two different idioms for achieving the same thing
can be found:

(a) what is done here, unconditionally add the postfunc, but then do an
early return if the USE flag is not set.

(b) use an intermediate variable which is empty by default but has an
override assignment conditional - this is e.g. found in u-boot.oeclass.

With my intended change of USE flag semantics, idiom (a) will break when
the USE flag has the value "0", since that string is considered true in
python. Instead of adding ugly »... and foo != "0"«, use this
opportunity to standardize on idiom (b).

I believe that is also slightly more correct, since it avoids
potentially confusing "running postfunc do_install_crontab" in
do_install.log when the user knows he disabled USE_crontab.