classes/chrpath.oeclass: use True instead of "1"
The parser ensures this is a no-op, but using True serves as
documentation that this really is a boolean flag, and not something
where the string value is used. It also prepares for a possibly distant
future or alternate universe where the parser actually stores the Python
object True when meeting the token True.