lib/oelite/fetch/url.py: rework and reenable proxy logic
It seems slightly silly that one cannot set a {ftp,https}_proxy without
having an http_proxy set, and having http_proxy work as a default for
the other two is not a that huge a convenience - one can just as easily
set all three to the same value.

In any case, use of those variables vanished when we switched to using
wget. This also reinstates that. Fortunately, the subprocess module
allows us to easily define the environment for the child, and "man wget"
says that it use the env variables {http,https,ftp}_proxy, so it doesn't
really require much code.