oelite/item.py: make it hashable
To be able to put OEliteItem in a set or hashtable, implement a
trivial hash method consistent with the __eq__ method.

While at it, refactor that slightly so that __eq__ and __hash__ use a
common helper, which also fixes the bug in the old __eq__ method that
referred to .item where it really meant .name (apparently, two
OEliteItems have never been compared for equality).