lib/oelite: prepare for getting rid of recipe_depends table
The current recipe_depends sqlite table makes it impossible for a
recipe to depend on two different versions of the same item. Moreover,
which version one ends up depending on is dependent on the order they
are encountered in the DEPENDS variable.

This commit introduces some very ugly code, but it only serves as a
check that we can implement the current logic using a couple of nested
hash tables hanging of the OEliteRecipe. When we eliminate the double
bookkeeping, everything should become much prettier, and we'll then
easily be able to change the logic to allow depending on multiple
versions of the same item.