rmwork: use proper count for remaining_tasks
In my reimplementation of rmwork, I'm going to rely on
recipe.remaining_tasks for detecting when there's no more tasks
belonging to a recipe, so it has to accurately reflect how many tasks
we're actually going to do for that recipe - not just how many task
functions the recipe metadata has. While at it, might as well report the
total number of tasks.

(It used to be used only for prioritizing among tasks belonging to
recipes with the same number of (recursive) dependants, so it didn't
matter too much what the original value was.)