ncurses: don't put binconfig script in separate package
It's silly to put the binconfig script into its own package; that just
forces anybody depending on ncurses to also depend on that
package. It should go into the ${PN}-dev package - there's already a
failed attempt at doing that, but the auto-package-utils packages come
before the default ones, so that line has no effect.

Instead, inherit the binconfig class which both sets the proper FILES_*
variable as well as ensuring binconfig-fixups get run for anything
depending on ncurses.