oelite/fetch/url.py: fix permissions of generated files
Downloaded files used to be created with mode 0o644 (modified by umask
as usual), but since we now let tempfile.mkstemp do the open(O_CREAT),
we get 0o600. This is a problem at least for mirroring where oe-lite
runs as one user but the webserver runs as another. The simplest way to
fix it up seems to be an explicit fchmod.