delete __mtimes when they're no longer needed
The __mtimes variable contains a list of typically ~60 tuples; the
deepcopying of that variable alone is responsible for about 1 second
of the ~25 seconds metadata hashing takes for oe bake world. It also
uses quite a lot of memory (a few MB).

Since we never use __mtimes after we've loaded and/or written the
cached recipe, might as well delete it.